+++ changelog.md
@@ -1,7 +1,11 @@
 ## Upcoming
-* Added back Absorption Armor, Armor rework, regen absorption, Bows, Piercing Damage and Misc Stats
+* Added back Absorption Armor, Armor rework, regen absorption, Bows, Snowballs, Piercing Damage and Misc Stats
   * Absorption armor is now enabled by default (for now)
   * Bows feature now includes No Arrow Invincibility Frames 
+  * Snowballs
+    * Slightly increased freeze time
+    * Added 4 ticks cooldown
+    * Now stack up to 64 with an integrated data pack thanks to InsaneLib's Item Components feature
   * Misc Stats is not complete yet
 * Ported Attack Speed Based Invincibility from ISO